527Project-master/requirements.txt
loguru
mysql-connector-python
pandas
redshift-connector
flask